# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Url-Shortener Application REST-API scheme",
        "title": "Golang Url-Shortener",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/urls": {
            "get": {
                "description": "Получить список всех ссылок с фильтрацией по user_id и пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Получить список ссылок",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список ссылок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_urls_transport_http.UrlDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать новую короткую ссылку для указанного URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Создать короткую ссылку",
                "parameters": [
                    {
                        "description": "CreateUrl тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_urls_transport_http.CreateUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданная ссылка",
                        "schema": {
                            "$ref": "#/definitions/internal_features_urls_transport_http.CreateUrlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/urls/{id}": {
            "get": {
                "description": "Получить информацию о конкретной короткой ссылке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Получить ссылку по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о ссылке",
                        "schema": {
                            "$ref": "#/definitions/internal_features_urls_transport_http.GetURLResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить короткую ссылку по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Удалить ссылку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ссылка удалена"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Сгенерировать новый короткий код для существующей ссылки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Обновить короткую ссылку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновлённая ссылка",
                        "schema": {
                            "$ref": "#/definitions/internal_features_urls_transport_http.PatchLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_IwantHappiness_url-shortener_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "user with id='42': not found"
                },
                "message": {
                    "type": "string",
                    "example": "failed to get user"
                }
            }
        },
...
        "internal_features_urls_transport_http.CreateUrlRequest": {
            "type": "object",
            "required": [
                "url",
                "user_id"
            ],
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://example.com"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_features_urls_transport_http.CreateUrlResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-06-28T15:04:05Z07:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "original_url": {
                    "type": "string",
                    "example": "https://example.com"
                },
                "short_url": {
                    "type": "string",
                    "example": "abc123"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_features_urls_transport_http.GetURLResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-06-28T15:04:05Z07:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "original_url": {
                    "type": "string",
                    "example": "https://example.com"
                },
                "short_url": {
                    "type": "string",
                    "example": "abc123"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_features_urls_transport_http.PatchLinkResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-06-28T15:04:05Z07:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "original_url": {
                    "type": "string",
                    "example": "https://example.com"
                },
                "short_url": {
                    "type": "string",
                    "example": "abc123"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_features_urls_transport_http.UrlDTOResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-06-28T15:04:05Z07:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "original_url": {
                    "type": "string",
                    "example": "https://example.com"
                },
                "short_url": {
                    "type": "string",
                    "example": "abc123"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}